requirements.txt
blinker==1.6.2
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
fastai==2.7.12
fastcore==1.5.29
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.12.0
Flask==2.3.1
fonttools==4.39.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
murmurhash==1.0.9
networkx==3.1
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
pathy==0.10.1
Pillow==9.5.0
preshed==3.0.8
pydantic==1.10.7
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.29.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
smart-open==6.3.0
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
sympy==1.11.1
thinc==8.1.9
threadpoolctl==3.1.0
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wasabi==1.1.1
Werkzeug==2.3.2